Mirkat 22 fix plot chart (#10)
* Remove the plot node form the constructor to avoid confusion

* Change scheme or the plot node

* Test that the plot node works properly

* Moving nodes

The nodes I used them in several parts, and looking for them in agent is messy, I move them to their file so we can focus on its configuration better.
I also moved GraphState

* Adding the instructions and solidifying the tests for Plot

* The test in pycharm mess up with the path, so I had to do this.

* Adding the dummy file of a successful Plot run_model

* typo on instructions

* Adding the plot-compleatness

The plot node had some changes on the way it display the results and the main issue was the extraction of the result from the plot. Probably there will be even more changes.

* Preparing the completeness

* Update instructions.yaml

* Change from binary file to image path in plot

* Update test_plot_node.py

* Get the plots

Use ****FINAL_RESPONSE**** to delimit when it finished and it is ready to plot.

Also uses stremlit.image to plot the picture
after ploting it save (root for now)

* Update .gitignore

* Update nodes.py

* Update node_plot.py